requirements.txt
Flask
requests
huggingface_hub
clerk-sdk-python